Se agrego la logica de recorte de espectros al paso Plate Segmentation.
